backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
blinker==1.7.0
certifi==2023.11.17
chardet==5.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.7
Django==4.2.11
djangorestframework==3.14.0
django-jazzmin==3.0.1
gunicorn==20.1.0
idna==3.6
Jinja2==3.1.2
kombu==5.3.5
Markdown==3.4.1
MarkupSafe==2.1.5
packaging==24.0
pillow==11.2.1
psutil==5.9.8
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.0.7
whitenoise==6.5.0
python-decouple==3.8
mysqlclient==2.2.7
